# back-end/public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Education website with Swagger",
        "description": "Getting started with Laravel and Swagger",
        "termsOfService": "",
        "contact": {
            "email": "name@example.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/cart_present/destroyCartPresent/{id}": {
            "delete": {
                "description": "Return cart present's information",
                "responses": {
                    "200": {
                        "description": "Delete cart present successfully",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "user_id": {
                                    "type": "integer"
                                },
                                "present_id": {
                                    "type": "integer"
                                },
                                "status": {
                                    "type": "integer"
                                },
                                "created_at": {
                                    "type": "timestamp"
                                },
                                "updated_at": {
                                    "type": "timestamp"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid id !!!"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/cart_present/changeStatusAdmin/{id}": {
            "get": {
                "description": "Return cart present's information",
                "responses": {
                    "200": {
                        "description": "Delete cart present successfully",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "user_id": {
                                    "type": "integer"
                                },
                                "present_id": {
                                    "type": "integer"
                                },
                                "status": {
                                    "type": "integer"
                                },
                                "created_at": {
                                    "type": "timestamp"
                                },
                                "updated_at": {
                                    "type": "timestamp"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid id !!!"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/cart_present/changeStatusClient/{id}": {
            "get": {
                "description": "Return cart present's information",
                "responses": {
                    "200": {
                        "description": "Delete cart present successfully",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "user_id": {
                                    "type": "integer"
                                },
                                "present_id": {
                                    "type": "integer"
                                },
                                "status": {
                                    "type": "integer"
                                },
                                "created_at": {
                                    "type": "timestamp"
                                },
                                "updated_at": {
                                    "type": "timestamp"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid id !!!"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
...
}